fix "ID [Sleep] already registered"
Running the sleep client throws the following error:

terminate called after throwing an instance of 'BT::BehaviorTreeException'
  what():  ID [Sleep] already registered
[ros2run]: Aborted

This error is caused by trying to register the Sleep Action with the Name "Sleep", which is likely due to the fact that behaviortree CPP already registers this ID as seen here:

https://github.com/BehaviorTree/BehaviorTree.CPP/blob/74c7dd7d5043a992426110b23eedca0fc04db2ae/src/bt_factory.cpp#L82

One Workaround to keep the Examples in this repo working is to rename the ID "Sleep" to "SleepAction". 
This fixes the issue and one can then run succesfully the sleep_server and then the sleep_client.